Make a commonly used method as an extension
