+++ Example/Example/ViewController.swift
@@ -40,7 +40,7 @@ class ViewController: UIViewController {
         do {
             try pixleeCredentials = PixleeCredentials.create()
         } catch{
-            showPopup(message: error.localizedDescription)
+            self.showPopup(message: error.localizedDescription)
         }
         
         initClient()
@@ -221,20 +221,5 @@ extension ViewController: PXLPhotoProductDelegate {
     
     public func onProductClicked(product: PXLProduct) {
         print("Pruduct: \(product.identifier) clicked")
-    }
-}
-
-
-// MARK: - Show Popup
-extension ViewController {
-    func showPopup(message:String) {
-        let alert = UIAlertController(title: "No credential file", message: message, preferredStyle: .alert)
-        let action = UIAlertAction(title: "Close", style: .default) { (action) in
-            
-        }
-        
-        alert.addAction(action)
-        
-        present(alert, animated: true, completion: nil)
     }
 }
